changed directory to %APPDATA%.mango and fixed bugs when really far out (2 billion blocks)

+++ Mango/app/player.h
@@ -15,29 +15,29 @@ public:
 	void OnUpdate() override { Update(); }
 
 	// getters/setters
-	void SetViewangle(const glm::vec3& viewangle) { m_viewangle = viewangle; }
-	glm::vec3 GetViewangle() const { return m_viewangle; }
-	void SetPosition(const glm::vec3& position) { m_position = position; }
-	glm::vec3 GetPosition() const { return m_position; }
-	void SetVelocity(const glm::vec3& velocity) { m_velocity = velocity; }
-	glm::vec3 GetVelocity() const { return m_velocity; }
-	void SetAcceleration(const glm::vec3& acceleration) { m_acceleration = acceleration; }
-	glm::vec3 GetAcceleration() const { return m_acceleration; }
-	void SetPrevViewangle(const glm::vec3& viewangle) { m_prev_viewangle = viewangle; }
-	glm::vec3 GetPrevViewangle() const { return m_prev_viewangle; }
-	void SetPrevPosition(const glm::vec3& position) { m_prev_position = position; }
-	glm::vec3 GetPrevPosition() const { return m_prev_position; }
-	void SetPrevVelocity(const glm::vec3& velocity) { m_prev_velocity = velocity; }
-	glm::vec3 GetPrevVelocity() const { return m_prev_velocity; }
-	void SetPrevAcceleration(const glm::vec3& acceleration) { m_prev_acceleration = acceleration; }
-	glm::vec3 GetPrevAcceleration() const { return m_prev_acceleration; }
+	void SetViewangle(const glm::dvec3& viewangle) { m_viewangle = viewangle; }
+	glm::dvec3 GetViewangle() const { return m_viewangle; }
+	void SetPosition(const glm::dvec3& position) { m_position = position; }
+	glm::dvec3 GetPosition() const { return m_position; }
+	void SetVelocity(const glm::dvec3& velocity) { m_velocity = velocity; }
+	glm::dvec3 GetVelocity() const { return m_velocity; }
+	void SetAcceleration(const glm::dvec3& acceleration) { m_acceleration = acceleration; }
+	glm::dvec3 GetAcceleration() const { return m_acceleration; }
+	void SetPrevViewangle(const glm::dvec3& viewangle) { m_prev_viewangle = viewangle; }
+	glm::dvec3 GetPrevViewangle() const { return m_prev_viewangle; }
+	void SetPrevPosition(const glm::dvec3& position) { m_prev_position = position; }
+	glm::dvec3 GetPrevPosition() const { return m_prev_position; }
+	void SetPrevVelocity(const glm::dvec3& velocity) { m_prev_velocity = velocity; }
+	glm::dvec3 GetPrevVelocity() const { return m_prev_velocity; }
+	void SetPrevAcceleration(const glm::dvec3& acceleration) { m_prev_acceleration = acceleration; }
+	glm::dvec3 GetPrevAcceleration() const { return m_prev_acceleration; }
 
-	float GetMaxSpeed() const { return m_max_speed; }
-	void SetMaxSpeed(float speed) { m_max_speed = speed; }
-	float GetFriction() const { return m_friction; }
-	void SetFriction(float friction) { m_friction = friction; }
-	float GetAccelSpeed() const { return m_accel_speed; }
-	void SetAccelSpeed(float accel_speed) { m_accel_speed = accel_speed; }
+	double GetMaxSpeed() const { return m_max_speed; }
+	void SetMaxSpeed(double speed) { m_max_speed = speed; }
+	double GetFriction() const { return m_friction; }
+	void SetFriction(double friction) { m_friction = friction; }
+	double GetAccelSpeed() const { return m_accel_speed; }
+	void SetAccelSpeed(double accel_speed) { m_accel_speed = accel_speed; }
 
 	bool IsOnGround() const { return m_on_ground; }
 
@@ -48,26 +48,26 @@ public:
 	void FrameUpdate(Mango::MangoCore* mango_core, float lerp);
 	void Update();
 
-	void SimulateMovement(glm::vec3& position, glm::vec3& velocity, glm::vec3 acceleration, bool& on_ground);
+	void SimulateMovement(glm::dvec3& position, glm::dvec3& velocity, glm::dvec3 acceleration, bool& on_ground);
 
 private:
 	Mango::Model m_model;
 	Mango::Shader* m_shader;
 
-	glm::vec3 m_viewangle = { 0.f, 0.f, 0.f },
-		m_prev_viewangle = { 0.f, 0.f, 0.f };
-	glm::vec3 m_position = { 0.f, 0.f, 0.f },
-		m_prev_position = { 0.f, 0.f, 0.f };
-	glm::vec3 m_velocity = { 0.f, 0.f, 0.f },
-		m_prev_velocity = { 0.f, 0.f, 0.f };
-	glm::vec3 m_acceleration = { 0.f, 0.f, 0.f },
-		m_prev_acceleration = { 0.f, 0.f, 0.f };
+	glm::dvec3 m_viewangle = { 0.0, 0.0, 0.0 },
+		m_prev_viewangle = { 0.0, 0.0, 0.0 };
+	glm::dvec3 m_position = { 0.0, 0.0, 0.0 },
+		m_prev_position = { 0.0, 0.0, 0.0 };
+	glm::dvec3 m_velocity = { 0.0, 0.0, 0.0 },
+		m_prev_velocity = { 0.0, 0.0, 0.0 };
+	glm::dvec3 m_acceleration = { 0.0, 0.0, 0.0 },
+		m_prev_acceleration = { 0.0, 0.0, 0.0 };
 
 	// units per second
-	float m_max_speed = 4.3f;
-	float m_friction = 16.f;
-	float m_accel_speed = 24.f;
-	float m_gravity = 16.f;
+	double m_max_speed = 4.3;
+	double m_friction = 16.0;
+	double m_accel_speed = 24.0;
+	double m_gravity = 16.0;
 
 	bool m_on_ground = false;
 };
